+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.4] - 2026-03-09
+
+### Changed
+- Rename instrument package to instrumentenum
+- Extended instrument enums with ResultType, ReagentType, entries in Ability
+- Fixed redis refresh validity state
+
 ## [1.1.3] - 2026-03-03
 
 ### Changed